Some small fixes

+++ clean.py
@@ -8,9 +8,11 @@ problemDirectories = [f.path for f in os.scandir(
 
 
 for problemDirectory in problemDirectories:
+    if problemDirectory[str(problemDirectory).rfind("/")+1:][0] == ".":
+        continue
     # Moving all the solution files out into the main Contest Directory
     os.rename(os.path.join(problemDirectory, "sol.cpp"), os.path.join(
-        contestDirectory, problemDirectory[-1]+".cpp"))
+        contestDirectory, problemDirectory+".cpp"))
     # Deleting all the Problem directories
     shutil.rmtree(problemDirectory)
